Updated imports of moved functions
